backend/requirements.txt
fastapi
uvicorn[standard]
motor
passlib[bcrypt]
PyJWT
httpx
celery
aiobreaker
redis
email-validator
